+++ pennylane_lightning/core/src/simulators/lightning_qubit/bindings/LQubitBindings.hpp
@@ -179,23 +179,25 @@ void registerBackendClassSpecificBindings(PyClass &pyclass) {
         .def("resetStateVector", &StateVectorT::resetStateVector)
         .def(
             "setBasisState",
-            [](StateVectorT &sv, const std::size_t index) {
-                sv.setBasisState(index);
+            [](StateVectorT &sv, const std::vector<std::size_t> &state,
+               const std::vector<std::size_t> &wires) {
+                sv.setBasisState(state, wires);
             },
-            "Create Basis State.")
+            "Set the state vector to a basis state.")
         .def(
             "setStateVector",
-            [](StateVectorT &sv, const std::vector<std::size_t> &indices,
-               const np_arr_c &state) {
+            [](StateVectorT &sv, const np_arr_c &state,
+               const std::vector<std::size_t> &wires) {
                 const auto buffer = state.request();
-                std::vector<ComplexT> state_in;
+                std::vector<ComplexT> state_vector;
                 if (buffer.size) {
                     const auto ptr = static_cast<const ComplexT *>(buffer.ptr);
-                    state_in = std::vector<ComplexT>{ptr, ptr + buffer.size};
+                    state_vector =
+                        std::vector<ComplexT>{ptr, ptr + buffer.size};
                 }
-                sv.setStateVector(indices, state_in);
+                sv.setStateVector(state_vector, wires);
             },
-            "Set State Vector with values and their corresponding indices")
+            "Set the state vector to the data contained in `state`.")
         .def(
             "getState",
             [](const StateVectorT &sv, np_arr_c &state) {