+++ pennylane_lightning/core/src/simulators/lightning_qubit/tests/Test_StateVectorLQubit.cpp
@@ -56,55 +56,54 @@ TEMPLATE_PRODUCT_TEST_CASE("StateVectorLQubit::setBasisState",
 
     using StateVectorT = TestType;
     using ComplexT = typename StateVectorT::ComplexT;
-    using PrecisionT = typename StateVectorT::PrecisionT;
 
     SECTION("setBasisState") {
 
-        const ComplexT init{1.0, PrecisionT{0.0}};
-        const ComplexT zero{PrecisionT{0.0}, PrecisionT{0.0}};
+        const ComplexT one{1.0};
+        const ComplexT zero{0.0};
 
         std::vector<ComplexT> init_state_zeros = {zero, zero, zero, zero,
                                                   zero, zero, zero, zero};
-        std::vector<ComplexT> expected_state_000 = {init, zero, zero, zero,
+        std::vector<ComplexT> expected_state_000 = {one,  zero, zero, zero,
                                                     zero, zero, zero, zero};
 
         StateVectorT sv(init_state_zeros.data(), init_state_zeros.size());
         sv.setBasisState({0}, {0});
         REQUIRE(sv.getDataVector() == approx(expected_state_000));
 
-        std::vector<ComplexT> expected_state_001 = {zero, init, zero, zero,
+        std::vector<ComplexT> expected_state_001 = {zero, one,  zero, zero,
                                                     zero, zero, zero, zero};
         sv.setBasisState({1}, {2});
         REQUIRE(sv.getDataVector() == approx(expected_state_001));
 
-        std::vector<ComplexT> expected_state_010 = {zero, zero, init, zero,
+        std::vector<ComplexT> expected_state_010 = {zero, zero, one,  zero,
                                                     zero, zero, zero, zero};
         sv.setBasisState({1}, {1});
         REQUIRE(sv.getDataVector() == approx(expected_state_010));
 
-        std::vector<ComplexT> expected_state_011 = {zero, zero, zero, init,
+        std::vector<ComplexT> expected_state_011 = {zero, zero, zero, one,
                                                     zero, zero, zero, zero};
         sv.setBasisState({1, 1}, {1, 2});
         REQUIRE(sv.getDataVector() == approx(expected_state_011));
 
         std::vector<ComplexT> expected_state_100 = {zero, zero, zero, zero,
-                                                    init, zero, zero, zero};
+                                                    one,  zero, zero, zero};
         sv.setBasisState({1}, {0});
         REQUIRE(sv.getDataVector() == approx(expected_state_100));
 
         std::vector<ComplexT> expected_state_101 = {zero, zero, zero, zero,
-                                                    zero, init, zero, zero};
+                                                    zero, one,  zero, zero};
 
         sv.setBasisState({1, 0, 1}, {0, 1, 2});
         REQUIRE(sv.getDataVector() == approx(expected_state_101));
 
         std::vector<ComplexT> expected_state_110 = {zero, zero, zero, zero,
-                                                    zero, zero, init, zero};
+                                                    zero, zero, one,  zero};
         sv.setBasisState({1, 1}, {0, 1});
         REQUIRE(sv.getDataVector() == approx(expected_state_110));
 
         std::vector<ComplexT> expected_state_111 = {zero, zero, zero, zero,
-                                                    zero, zero, zero, init};
+                                                    zero, zero, zero, one};
         sv.setBasisState({1, 1, 1}, {0, 1, 2});
         REQUIRE(sv.getDataVector() == approx(expected_state_111));
     }
@@ -117,64 +116,63 @@ TEMPLATE_PRODUCT_TEST_CASE("StateVectorLQubit::setStateVector",
 
     using StateVectorT = TestType;
     using ComplexT = typename StateVectorT::ComplexT;
-    using PrecisionT = typename StateVectorT::PrecisionT;
 
     SECTION("setStateVector") {
 
-        const ComplexT init{1.0, PrecisionT{0.0}};
-        const ComplexT zero{PrecisionT{0.0}, PrecisionT{0.0}};
-        const ComplexT half{0.5, PrecisionT{0.0}};
+        const ComplexT one{1.0};
+        const ComplexT zero{0.0};
+        const ComplexT half{0.5};
 
         std::vector<ComplexT> init_state_zeros = {zero, zero, zero, zero,
                                                   zero, zero, zero, zero};
-        std::vector<ComplexT> expected_state_000 = {init, zero, zero, zero,
+        std::vector<ComplexT> expected_state_000 = {one,  zero, zero, zero,
                                                     zero, zero, zero, zero};
 
         StateVectorT sv(init_state_zeros.data(), init_state_zeros.size());
-        sv.setStateVector({init, zero}, {0});
+        sv.setStateVector({one, zero}, {0});
         REQUIRE(sv.getDataVector() == approx(expected_state_000));
 
-        std::vector<ComplexT> expected_state_001 = {zero, init, zero, zero,
+        std::vector<ComplexT> expected_state_001 = {zero, one,  zero, zero,
                                                     zero, zero, zero, zero};
         sv.resetStateVector();
-        sv.setStateVector({zero, init}, {2});
+        sv.setStateVector({zero, one}, {2});
         REQUIRE(sv.getDataVector() == approx(expected_state_001));
 
-        std::vector<ComplexT> expected_state_010 = {zero, zero, init, zero,
+        std::vector<ComplexT> expected_state_010 = {zero, zero, one,  zero,
                                                     zero, zero, zero, zero};
 
         sv.resetStateVector();
-        sv.setStateVector({zero, init}, {1});
+        sv.setStateVector({zero, one}, {1});
         REQUIRE(sv.getDataVector() == approx(expected_state_010));
 
-        std::vector<ComplexT> expected_state_011 = {zero, zero, zero, init,
+        std::vector<ComplexT> expected_state_011 = {zero, zero, zero, one,
                                                     zero, zero, zero, zero};
         sv.resetStateVector();
-        sv.setStateVector({zero, zero, zero, init}, {1, 2});
+        sv.setStateVector({zero, zero, zero, one}, {1, 2});
         REQUIRE(sv.getDataVector() == approx(expected_state_011));
 
         std::vector<ComplexT> expected_state_100 = {zero, zero, zero, zero,
-                                                    init, zero, zero, zero};
+                                                    one,  zero, zero, zero};
         sv.resetStateVector();
-        sv.setStateVector({zero, init}, {0});
+        sv.setStateVector({zero, one}, {0});
         REQUIRE(sv.getDataVector() == approx(expected_state_100));
 
         std::vector<ComplexT> expected_state_101 = {zero, zero, zero, zero,
-                                                    zero, init, zero, zero};
+                                                    zero, one,  zero, zero};
         sv.resetStateVector();
-        sv.setStateVector({zero, zero, zero, init}, {0, 2});
+        sv.setStateVector({zero, zero, zero, one}, {0, 2});
         REQUIRE(sv.getDataVector() == approx(expected_state_101));
 
         std::vector<ComplexT> expected_state_110 = {zero, zero, zero, zero,
-                                                    zero, zero, init, zero};
+                                                    zero, zero, one,  zero};
         sv.resetStateVector();
-        sv.setStateVector({zero, zero, zero, init}, {0, 1});
+        sv.setStateVector({zero, zero, zero, one}, {0, 1});
         REQUIRE(sv.getDataVector() == approx(expected_state_110));
 
         std::vector<ComplexT> expected_state_111 = {zero, zero, zero, zero,
-                                                    zero, zero, zero, init};
+                                                    zero, zero, zero, one};
         sv.resetStateVector();
-        sv.setStateVector({zero, zero, zero, zero, zero, zero, zero, init},
+        sv.setStateVector({zero, zero, zero, zero, zero, zero, zero, one},
                           {0, 1, 2});
         REQUIRE(sv.getDataVector() == approx(expected_state_111));
     }